+++ minos/aggregate/__init__.py
@@ -1,6 +1,6 @@
 __author__ = """Clariteia Devs"""
 __email__ = "devs@clariteia.com"
-__version__ = "0.2.3"
+__version__ = "0.2.4"
 
 from .contextvars import (
     IS_REPOSITORY_SERIALIZATION_CONTEXT_VAR,