+++ minos/aggregate/models/collections.py
@@ -17,8 +17,12 @@ from typing import (
 )
 
 from minos.common import (
+    DataDecoder,
+    DataEncoder,
     DeclarativeModel,
     ModelType,
+    SchemaDecoder,
+    SchemaEncoder,
 )
 
 from .actions import (
@@ -41,14 +45,14 @@ class IncrementalSet(DeclarativeModel, MutableSet, Generic[T]):
         super().__init__(data, *args, **kwargs)
 
     def add(self, value_object: T) -> None:
-        """Add an value object.
+        """Add a value object.
         :param value_object: The value object to be added.
         :return: This method does not return anything.
         """
         self.data.add(value_object)
 
     def discard(self, value_object: T) -> None:
-        """Remove an value object.
+        """Remove a value object.
         :param value_object: The value object to be added.
         :return: This method does not return anything.
         """
@@ -83,6 +87,51 @@ class IncrementalSet(DeclarativeModel, MutableSet, Generic[T]):
         """
         args = get_args(self.type_hints["data"])
         return args[0]
+
+    # noinspection PyMethodParameters
+    @classmethod
+    def encode_schema(cls, encoder: SchemaEncoder, target: Any, **kwargs) -> Any:
+        """Encode schema with the given encoder.
+
+        :param encoder: The encoder instance.
+        :param target: An optional pre-encoded schema.
+        :return: The encoded schema of the instance.
+        """
+        schema = encoder.build(target.type_hints["data"], **kwargs)
+        return schema | {"logicalType": cls.classname}
+
+    @classmethod
+    def decode_schema(cls, decoder: SchemaDecoder, target: Any, **kwargs) -> ModelType:
+        """Decode schema with the given encoder.
+
+        :param decoder: The decoder instance.
+        :param target: The schema to be decoded.
+        :return: The decoded schema as a type.
+        """
+        decoded = decoder.build(target, **kwargs)
+        return ModelType.from_model(cls[get_args(decoded)[-1]])
+
+    @staticmethod
+    def encode_data(encoder: DataEncoder, target: Any, **kwargs) -> Any:
+        """Encode data with the given encoder.
+
+        :param encoder: The encoder instance.
+        :param target: An optional pre-encoded data.
+        :return: The encoded data of the instance.
+        """
+        return encoder.build(target["data"], **kwargs)
+
+    @classmethod
+    def decode_data(cls, decoder: DataDecoder, target: Any, type_: ModelType, **kwargs) -> IncrementalSet:
+        """Decode data with the given decoder.
+
+        :param decoder: The decoder instance.
+        :param target: The data to be decoded.
+        :param type_: The data type.
+        :return: A decoded instance.
+        """
+        decoded = decoder.build(target, type_.type_hints["data"], **kwargs)
+        return cls(decoded, additional_type_hints=type_.type_hints)
 
 
 IncrementalSetDiffEntry = ModelType.build("SetDiffEntry", {"action": Action, "entity": Any})