+++ HISTORY.md
@@ -33,10 +33,15 @@
 * Add `ModelRefResolver` to resolve `ModelRef` instances in batch.
 * Fix bug that published *Domain Events* when non-top-level transactions are committed.
 
-
 ## 0.2.3 (2022-01-10)
 
 * Remove the `data` wrapper from the `ModelRef`'s avro serialization.
 * Add support for search by `ModelRef` instances on the `PostgreSqlSnapshot`.
 * Fix bug related with the `IS_REPOSITORY_SERIALIZATION_CONTEXT_VAR`, used to store on the repository only the references to another aggregates.
 * Fix bug related with concurrent enabling of the `uuid-ossp` extension from `postgres`.
+
+## 0.2.4 (2022-01-11)
+
+* Remove the `data` wrapper from the `IncrementalSet`'s avro serialization.
+* Remove the `data` wrapper from the `ValueObjectSet`'s avro serialization.
+* Remove the `data` wrapper from the `EntitySet`'s avro serialization.